Fuel Gauge Addition
A fuel gauge was added to the bottom left hand side of the screen. This
game object follows the 2D camera around so it stays put. An arbitrary
fuel amount was added to the Lunar Lander script, and it was linked to a
fuel gauge class. As the user uses thrust, the fuel will deplete and a
visual representation is updated. Once the fuel gauge is empty, then the
thrusters halt operation.