Fuel, HP, Win Landing Area
I adjusted the fuel gauge, and added an HP gauge. The fuel is controlled
by the landercontroller script, where a fuel amount is defined and
adjusted during thruster use. The HP amount will change when the lander
collides with an collider and the velocity of the lander exceeds a set
threshold. A fuelgauge and HP script has been added to work with the
lander script.
The landing area is just some code added to the lander controller script
that recognizes when the lander is continuously colliding with the
collider tagged "winArea".